Finish adding track action events to RealTime
