+++ MercadoPagoPXTracking/MPTTrackType.swift
@@ -1,0 +1,13 @@
+//
+//  MPTTrackType.swift
+//  MercadoPagoPXTracking
+//
+//  Created by Eden Torres on 3/12/18.
+//  Copyright © 2018 Mercado Pago. All rights reserved.
+//
+
+import Foundation
+internal enum MPTTrackType: String {
+    case ACTION = "action"
+    case SCREENVIEW = "screenview"
+}
