+++ MercadoPagoPXTracking/MPTrackInfo.swift
@@ -1,0 +1,12 @@
+//
+//  MPTrackInfo.swift
+//  MercadoPagoPXTracking
+//
+//  Created by Eden Torres on 3/12/18.
+//  Copyright © 2018 Mercado Pago. All rights reserved.
+//
+
+import Foundation
+protocol MPTrackInfo {
+    func toJSON() -> [String: Any]
+}
