+++ MercadoPagoPXTracking/TrackingServices.swift
@@ -19,6 +19,11 @@ class TrackingServices: NSObject {
             requesturl += "?" + params!
         }
         let finalURL: NSURL = NSURL(string: requesturl)!
+
+//        #if DEBUG
+//            print("\n--REQUEST_URL: \(finalURL)")
+//        #endif
+
         let request: NSMutableURLRequest
         request = NSMutableURLRequest(url: finalURL as URL)
         request.url = finalURL as URL
@@ -30,18 +35,20 @@ class TrackingServices: NSObject {
             }
         }
         if let body = body {
+//            #if DEBUG
+//                print("--REQUEST_BODY: \(body as! NSString)")
+//            #endif
             request.httpBody = body.data(using: String.Encoding.utf8)
         }
         let session = URLSession.shared
         let task = session.dataTask(with: request as URLRequest, completionHandler: {data, response, error -> Void in
-            UIApplication.shared.isNetworkActivityIndicatorVisible = false
             if error == nil {
                 do {
                     if let response = response as? HTTPURLResponse {
                         if response.statusCode == TrackingServices.STATUS_OK {
                             if let data = data {
                                 let responseJson = try JSONSerialization.jsonObject(with: data,
-                                                                                    options:JSONSerialization.ReadingOptions.allowFragments)
+                                                                                options:JSONSerialization.ReadingOptions.allowFragments)
                                 success(responseJson as Any)
 
                             }else {
